refactoring WS ref related class names
